requirements.txt
dj-database-url==0.5.0
django==2.2.9
django-cors-headers==3.2.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
django-allauth==0.36.0
django-rest-auth==0.9.3
django-anymail==7.0.0
django-environ==0.4.4
drfdocs
gunicorn==20.0.4
psycopg2-binary==2.8.4
pylint==2.4.4
pytz==2019.3
whitenoise==5.0.1
hashids
django-heroku
django-redis==4.9.0
django-q==1.0.2
requests==2.22.0
sentry-sdk
django-q[sentry]
ipython